[fix] Add check to overwrite folders Check that folders exist before trying to remove them for the --overwrite flag.

+++ miracl/flow/ace_test_interface_abstract.py
@@ -615,7 +615,8 @@ class CheckOverwriteFlag:
         )
 
         for folder in folder_locations:
-            shutil.rmtree(folder)
+            if folder.is_dir():
+                shutil.rmtree(folder)
     
     @staticmethod
     def _get_dirs(
